Split `coconext.triggers` into multiple modules
